+++ eslint.config.js
@@ -1,0 +1,40 @@
+import eslint from '@eslint/js';
+import tseslint from 'typescript-eslint';
+import tsParser from '@typescript-eslint/parser';
+
+export default tseslint.config(
+  {
+    ignores: ['dist/**'],
+  },
+  {
+    languageOptions: {
+      parser: tsParser,
+      ecmaVersion: 2022,
+      sourceType: 'module',
+      parserOptions: {
+        projectService: true,
+        tsconfigRootDir: import.meta.dirname,
+      },
+    },
+    extends: [
+      eslint.configs.recommended,
+      tseslint.configs.recommended, // could use RecommendedTypedChecked
+    ],
+    rules: {
+      'quotes': ['error', 'single'],
+      'linebreak-style': ['error', 'unix'],
+      'semi': ['error', 'always'],
+      'comma-dangle': ['error', 'always-multiline'],
+      'dot-notation': 'error',
+      'eqeqeq': ['error', 'smart'],
+      'curly': ['error', 'all'],
+      'brace-style': ['error'],
+      'prefer-arrow-callback': 'warn',
+      'max-len': ['warn', 160],
+      'object-curly-spacing': ['error', 'always'],
+      'no-use-before-define': 'off',
+      '@typescript-eslint/no-use-before-define': ['error', { 'classes': false, 'enums': false }],
+      '@typescript-eslint/no-unused-vars': ['error', { 'caughtErrors': 'none' }],
+    },
+  },
+);