+++ src/interfaces/tv-app.interface.ts
@@ -1,5 +1,3 @@
-import {Service} from 'homebridge';
-
 export interface TVApp {
   url: string;
   isunInstalled: boolean;
@@ -7,5 +5,4 @@ export interface TVApp {
   id: string;
   urlType?: number|string;
   storeType: number;
-  service?: Service;
 }