+++ src/platform.ts
@@ -1,26 +1,31 @@
-import {API, DynamicPlatformPlugin, PlatformAccessory, PlatformConfig, Service, Characteristic, Categories, Logging} from 'homebridge';
+import {
+  API,
+  PlatformConfig,
+  Service,
+  Characteristic,
+  Categories,
+  Logging,
+  IndependentPlatformPlugin,
+} from 'homebridge';
 
 import { PLUGIN_NAME } from './settings.js';
 import { HiSenseTVAccessory } from './platformAccessory.js';
-import {DeviceConfig} from './interfaces/device-config.interface.js';
-import {validateHomeKitName} from './utils/validateHomeKitName.function.js';
+import { DeviceConfig } from './interfaces/device-config.interface.js';
+import { validateHomeKitName } from './utils/validateHomeKitName.function.js';
 
 /**
  * HomebridgePlatform
  * This class is the main constructor for your plugin, this is where you should
  * parse the user config and discover/register accessories with Homebridge.
  */
-export class HiSenseTVPlatform implements DynamicPlatformPlugin {
+export class HiSenseTVPlatform implements IndependentPlatformPlugin {
   public readonly Service: typeof Service;
   public readonly Characteristic: typeof Characteristic;
 
-  // this is used to track restored cached accessories
-  public readonly accessories: PlatformAccessory[] = [];
-
   constructor(
-    public readonly log: Logging,
-    public readonly config: PlatformConfig,
-    public readonly api: API,
+      public readonly log: Logging,
+      public readonly config: PlatformConfig,
+      public readonly api: API,
   ) {
     this.Service = this.api.hap.Service;
     this.Characteristic = this.api.hap.Characteristic;
@@ -33,10 +38,15 @@ export class HiSenseTVPlatform implements DynamicPlatformPlugin {
       log.success = log.info;
     }
 
+    // comment from homebridge-team
     // When this event is fired it means Homebridge has restored all cached accessories from disk.
     // Dynamic Platform plugins should only register new accessories after this event was fired,
     // in order to ensure they weren't added to homebridge already. This event can also be used
     // to start discovery of new accessories.
+
+    // comment from maintainer
+    // We don't have a DynamicPlatformPlugin, so we don't necessarily have to wait for this event,
+    // but we will do it anyway to be safe.
     this.api.on('didFinishLaunching', () => {
       log.debug('Executed didFinishLaunching callback');
       // run the method to discover / register your devices as accessories
@@ -44,21 +54,6 @@ export class HiSenseTVPlatform implements DynamicPlatformPlugin {
     });
   }
 
-  /**
-   * This function is invoked when homebridge restores cached accessories from disk at startup.
-   * It should be used to setup event handlers for characteristics and update respective values.
-   */
-  configureAccessory(accessory: PlatformAccessory) {
-    this.log.info('Loading accessory from cache:', accessory.displayName);
-
-    // add the restored accessory to the accessories cache so we can track if it has already been registered
-    this.accessories.push(accessory);
-  }
-
-  /**
-   * As we are only publishing external accessories, they don't get stored in the cache.
-   * So only the else is important here.
-   */
   discoverDevices() {
     const configDevices = this.config.devices as DeviceConfig[] || [];
 
@@ -68,50 +63,29 @@ export class HiSenseTVPlatform implements DynamicPlatformPlugin {
       // generate a unique id for the accessory this should be generated from
       // something globally unique, but constant, for example, the device serial
       // number or MAC address
-      const uuid = this.api.hap.uuid.generate(device.id+device.name);
+      const uuid = this.api.hap.uuid.generate(`homebridge-hisense-tv:${device.id}:${device.macaddress}`);
 
-      // see if an accessory with the same uuid has already been registered and restored from
-      // the cached devices we stored in the `configureAccessory` method above
-      const existingAccessory = this.accessories.find(accessory => accessory.UUID === uuid);
+      // the accessory does not yet exist, so we need to create it
+      this.log.info('Adding new accessory:', device.name);
 
-      if (existingAccessory) {
-        // the accessory already exists
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName);
+      // create a new accessory
+      const accessory = new this.api.platformAccessory(validateHomeKitName(device.name), uuid, Categories.TELEVISION);
 
-        // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
-        // existingAccessory.context.device = device;
-        // this.api.updatePlatformAccessories([existingAccessory]);
+      // store a copy of the device object in the `accessory.context`
+      // the `context` property can be used to store any data about the accessory you may need
+      accessory.context.device = device;
+      accessory.context.macaddress = this.config.macaddress;
 
-        // create the accessory handler for the restored accessory
-        // this is imported from `platformAccessory.ts`
-        new HiSenseTVAccessory(this, existingAccessory);
-
-        // update accessory cache with any changes to the accessory details and information
-        //this.api.updatePlatformAccessories([existingAccessory]);
-
-        // it is possible to remove platform accessories at any time using `api.unregisterPlatformAccessories`, eg.:
-        // remove platform accessories when no longer present
-        //this.api.unregisterPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [existingAccessory]);
-        //this.log.info('Removing existing accessory from cache:', existingAccessory.displayName);
-      } else {
-        // the accessory does not yet exist, so we need to create it
-        this.log.info('Adding new accessory:', device.name);
-
-        // create a new accessory
-        const accessory = new this.api.platformAccessory(validateHomeKitName(device.name), uuid, Categories.TELEVISION);
-
-        // store a copy of the device object in the `accessory.context`
-        // the `context` property can be used to store any data about the accessory you may need
-        accessory.context.device = device;
-        accessory.context.macaddress = this.config.macaddress;
-
-        // create the accessory handler for the newly create accessory
-        // this is imported from `platformAccessory.ts`
-        new HiSenseTVAccessory(this, accessory);
-
-        // link the accessory to your platform
-        this.api.publishExternalAccessories(PLUGIN_NAME, [accessory]);
-        // this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
+      // create the accessory handler for the newly create accessory
+      // this is imported from `platformAccessory.ts`
+      this.log.info(`Waiting for TV data before publishing: ${device.name}`);
+      try {
+        new HiSenseTVAccessory(this, accessory, () => {
+          this.log.info(`TV data ready, publishing to HomeKit: ${device.name}`);
+          this.api.publishExternalAccessories(PLUGIN_NAME, [accessory]);
+        });
+      } catch (e: unknown) {
+        this.log.error(`Failed to initialize TV "${device.name}": ${(e as Error).message}`);
       }
     }
   }